develop stage --snapshot nightly 6 april 2016 broadcast rcvr  changes

+++ app/src/main/java/com/george/autorunpro/AlarmManagerBroadcastReceiver.java
@@ -1,10 +1,16 @@
 package com.george.autorunpro;
 
 import java.util.Calendar;
+import java.util.List;
+import java.util.SortedMap;
+import java.util.TreeMap;
+
 import android.app.ActivityManager;
 import android.app.Notification;
 import android.app.NotificationManager;
 import android.app.PendingIntent;
+import android.app.usage.UsageStats;
+import android.app.usage.UsageStatsManager;
 import android.content.BroadcastReceiver;
 import android.content.ContentValues;
 import android.content.Context;
@@ -14,6 +20,7 @@ import android.content.pm.ApplicationInfo;
 import android.content.pm.PackageManager;
 import android.database.Cursor;
 import android.graphics.Camera;
+import android.os.Build;
 import android.os.Bundle;
 import android.os.PowerManager;
 import android.util.Log;
@@ -165,25 +172,77 @@ public class AlarmManagerBroadcastReceiver extends BroadcastReceiver {
     void startapplication(Context context) {
         PackageManager packm = context.getPackageManager();
         Intent LaunchIntent = packm.getLaunchIntentForPackage(packageName);
-        Notification_Service notification_service = new Notification_Service(context,"Starting App","AutoRun Pro has started "+getTitle(context,packageName)+" for you");
-        notification_service.showNotification();
         if (LaunchIntent != null) {
+            Notification_Service notification_service = new Notification_Service(context,"AutoRun Pro","Started "+getTitle(context,packageName)+" for you");
+            notification_service.showNotification();
             context.startActivity(LaunchIntent);
         } else {
-            Toast.makeText(context, "Application  not found", Toast.LENGTH_SHORT).show();
+            Notification_Service notification_service = new Notification_Service(context,"AutoRun Pro","Error!, cannot start "+getTitle(context,packageName)+"");
+            notification_service.showNotification();
         }
     }
 
     void stopapplication(Context context){
-
-        Intent startMain = new Intent(Intent.ACTION_MAIN);
-        startMain.addCategory(Intent.CATEGORY_HOME);
-        startMain.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-        context.startActivity(startMain);
+       //checking app is in foreground,then only a stop is needed
+        //stop refers to minimising as android dont like killing apps
         am = (ActivityManager)context.getSystemService(Context.ACTIVITY_SERVICE);
-        am.killBackgroundProcesses(packageName);
-        Notification_Service notification_service = new Notification_Service(context,"Stopping App","AutoRun Pro has stopped "+getTitle(context,packageName)+" for you");
-        notification_service.showNotification();
+        List<ActivityManager.RunningAppProcessInfo> procInfos = am.getRunningAppProcesses();
+        Log.i("on reciever","in the stop application");
+        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.LOLLIPOP) {
+            Log.i("For less than lollipop","hhhh");
+            for(ActivityManager.RunningAppProcessInfo appProcess : procInfos) {
+                if (appProcess.importance == ActivityManager.RunningAppProcessInfo.IMPORTANCE_FOREGROUND) {
+                    if (appProcess.processName.equals(packageName)) {
+                        Intent startMain = new Intent(Intent.ACTION_MAIN);
+                        startMain.addCategory(Intent.CATEGORY_HOME);
+                        startMain.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+                        context.startActivity(startMain);
+                        am.killBackgroundProcesses(packageName);
+
+                        Notification_Service notification_service = new Notification_Service(context, "AutoRun Pro", "Stopped " + getTitle(context, packageName) + " for you");
+                        notification_service.showNotification();
+
+                    }
+                    //Log.i("Foreground App", appProcess.processName);
+                }
+            }
+
+            }else{
+            Log.i("For lollipop and above","hhh");
+            @SuppressWarnings("WrongConstant")
+            UsageStatsManager usm = (UsageStatsManager) context.getSystemService("usagestats");
+            long time = System.currentTimeMillis();
+            List<UsageStats> appList = usm.queryUsageStats(UsageStatsManager.INTERVAL_DAILY,
+                    time - 1000 * 1000, time);
+            if (appList != null && appList.size() > 0) {
+                SortedMap<Long, UsageStats> mySortedMap = new TreeMap<Long, UsageStats>();
+                for (UsageStats usageStats : appList) {
+                    mySortedMap.put(usageStats.getLastTimeUsed(),
+                            usageStats);
+                }
+                if (mySortedMap != null && !mySortedMap.isEmpty()) {
+                    Log.i("Package Found on top",mySortedMap.get(mySortedMap.lastKey()).getPackageName());
+                    if(packageName.equals(mySortedMap.get(mySortedMap.lastKey()).getPackageName())){
+
+                        Intent startMain = new Intent(Intent.ACTION_MAIN);
+                        startMain.addCategory(Intent.CATEGORY_HOME);
+                        startMain.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+                        context.startActivity(startMain);
+                        am.killBackgroundProcesses(packageName);
+
+                        Notification_Service notification_service = new Notification_Service(context, "AutoRun Pro", "Stopped " + getTitle(context, packageName) + " for you");
+                        notification_service.showNotification();
+                    }
+                }
+            }
+
+
+            }
+
+
+
+
+
 
     }
     void choose(Context context,int mode){
@@ -197,21 +256,53 @@ public class AlarmManagerBroadcastReceiver extends BroadcastReceiver {
                 stopapplication(context);
         }
         else{
-
+             String status = "";
              System.out.println("in the chooser");
              FunctionEngine functionEngine = new FunctionEngine(context);
-             if (funcname.equals("Wifi")){functionEngine.wifi(mode);}
 
-             else if(funcname.equals("Bluetooth")){functionEngine.bluetooth(mode);}
+             if (funcname.equals("Wifi")){
 
-             else if(funcname.equals("Vibrate Mode")){functionEngine.vibrate(mode);}
+                 if(functionEngine.wifi(mode)){
+                     if(mode == 0)status = "Enabled";
+                     else status = "Disabled";
+                 }else{
+                    status = "Error in enabling/disabling";
+                 }
+             }
 
-             else if(funcname.equals("Silent Mode")){functionEngine.silent(mode);}
+             else if(funcname.equals("Bluetooth")){
+                 if(functionEngine.bluetooth(mode)){
+                     if(mode == 0)status = "Enabled";
+                     else status = "Disabled";
+                 }
+                 else {
+                     status = "Error in enabling/disabling";
+                 }
 
+             }
 
+             else if(funcname.equals("Vibrate Mode")){
+                 if(functionEngine.vibrate(mode)){
+                     if(mode == 0)status = "Enabled";
+                     else status = "Disabled";
+                 }
+                 else{
+                     status = "Error in enabling/disabling";
+                 }
+             }
 
+             else if(funcname.equals("Silent Mode")){
+                 if(functionEngine.silent(mode)){
+                     if(mode == 0)status = "Enabled";
+                     else status = "Disabled";
+                 }
+                 else{
+                     status = "Error in enabling/disabling";
+                 }
+             }
 
-
+            Notification_Service notification_service = new Notification_Service(context,"AutoRun Pro",status+" "+funcname+ " for you");
+            notification_service.showNotification();
         }
 
     }
@@ -226,7 +317,6 @@ public class AlarmManagerBroadcastReceiver extends BroadcastReceiver {
         }
         catch (final PackageManager.NameNotFoundException e) { }
         final String title = (String) ((applicationInfo != null) ? packageManager.getApplicationLabel(applicationInfo) : "???");
-
         return title;
     }
